requirements/requirements.txt
Django==1.11.18
django-ckeditor==5.6.1
django-js-asset==1.1.0
Pillow==5.4.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.7
